feat: reorganizing packages
