+++ packages/audioplayers/lib/audioplayers.dart
@@ -1,9 +1,9 @@
-export 'package:audioplayers_platform_interface/api/audio_context_config.dart';
-export 'package:audioplayers_platform_interface/api/global_event.dart';
-export 'package:audioplayers_platform_interface/api/player_event.dart';
-export 'package:audioplayers_platform_interface/api/player_mode.dart';
-export 'package:audioplayers_platform_interface/api/player_state.dart';
-export 'package:audioplayers_platform_interface/api/release_mode.dart';
+export 'package:audioplayers_platform_interface/src/api/audio_context_config.dart';
+export 'package:audioplayers_platform_interface/src/api/global_event.dart';
+export 'package:audioplayers_platform_interface/src/api/player_event.dart';
+export 'package:audioplayers_platform_interface/src/api/player_mode.dart';
+export 'package:audioplayers_platform_interface/src/api/player_state.dart';
+export 'package:audioplayers_platform_interface/src/api/release_mode.dart';
 
 export 'src/audio_cache.dart';
 export 'src/audio_pool.dart';
