+++ packages/audioplayers/test/fake_audioplayers_platform.dart
@@ -1,14 +1,11 @@
 import 'dart:async';
 
-import 'package:audioplayers_platform_interface/api/audio_context_config.dart';
-import 'package:audioplayers_platform_interface/api/global_event.dart';
-import 'package:audioplayers_platform_interface/api/player_event.dart';
-import 'package:audioplayers_platform_interface/api/player_mode.dart';
-import 'package:audioplayers_platform_interface/api/release_mode.dart';
+import 'package:audioplayers/audioplayers.dart';
 import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
 import 'package:flutter/services.dart';
 
-class FakeGlobalAudioplayersPlatform extends GlobalPlatformInterface {
+class FakeGlobalAudioplayersPlatform
+    extends GlobalAudioplayersPlatformInterface {
   List<String> calls = <String>[];
   StreamController<GlobalEvent> eventStreamController =
       StreamController<GlobalEvent>.broadcast();
