+++ packages/audioplayers_platform_interface/lib/src/api/global_event.dart
