+++ packages/audioplayers_platform_interface/lib/src/api/player_state.dart
