+++ packages/audioplayers_platform_interface/lib/src/audioplayers_platform.dart
@@ -4,13 +4,13 @@ import 'dart:async';
 // ignore: unnecessary_import
 import 'dart:typed_data';
 
-import 'package:audioplayers_platform_interface/api/audio_context_config.dart';
-import 'package:audioplayers_platform_interface/api/player_event.dart';
-import 'package:audioplayers_platform_interface/api/player_mode.dart';
-import 'package:audioplayers_platform_interface/api/release_mode.dart';
-import 'package:audioplayers_platform_interface/audioplayers_platform_interface.dart';
-import 'package:audioplayers_platform_interface/map_extension.dart';
-import 'package:audioplayers_platform_interface/method_channel_extension.dart';
+import 'package:audioplayers_platform_interface/src/api/audio_context_config.dart';
+import 'package:audioplayers_platform_interface/src/api/player_event.dart';
+import 'package:audioplayers_platform_interface/src/api/player_mode.dart';
+import 'package:audioplayers_platform_interface/src/api/release_mode.dart';
+import 'package:audioplayers_platform_interface/src/audioplayers_platform_interface.dart';
+import 'package:audioplayers_platform_interface/src/map_extension.dart';
+import 'package:audioplayers_platform_interface/src/method_channel_extension.dart';
 import 'package:flutter/services.dart';
 
 class AudioplayersPlatform extends AudioplayersPlatformInterface
