+++ packages/audioplayers_platform_interface/lib/src/global_audioplayers_platform_interface.dart
@@ -1,6 +1,6 @@
-import 'package:audioplayers_platform_interface/api/audio_context_config.dart';
-import 'package:audioplayers_platform_interface/api/global_event.dart';
-import 'package:audioplayers_platform_interface/global_audioplayers_platform.dart';
+import 'package:audioplayers_platform_interface/src/api/audio_context_config.dart';
+import 'package:audioplayers_platform_interface/src/api/global_event.dart';
+import 'package:audioplayers_platform_interface/src/global_audioplayers_platform.dart';
 import 'package:meta/meta.dart';
 
 abstract class GlobalAudioplayersPlatformInterface
