+++ packages/audioplayers_platform_interface/lib/src/map_extension.dart
