+++ packages/audioplayers_platform_interface/lib/src/method_channel_extension.dart
